requirements.txt
numpy==1.24.3
Flask==2.2.2
youtube_transcript_api
newspaper3k